Implement recipe browser app and Docker deployment
Add typed data (materials, recipes, effects), search/filter/favorites,
Sheikah-styled recipe cards with dish images and grouped ingredients,
custom effect icon set, and a Docker/nginx setup for self-hosting.

+++ eslint.config.js
@@ -24,7 +24,7 @@ export default defineConfig([
     },
     rules: {
       "@stylistic/indent": ["error", 2],
-      "@stylistic/quotes": ["error", "single"],
+      "@stylistic/quotes": ["error", "double"],
     },
   },
 ]);
